pyproject.toml
[tool.poetry]
name = "django-shop"
version = "0.1.0"
description = ""
authors = ["Lee Tran <nglouie.dev@gmail.com>"]
readme = "README.md"
packages = [{ include = "django_shop" }]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
pillow = "^10.1.0"
celery = "^5.3.6"
flower = "^2.0.1"
stripe = "^7.10.0"
weasyprint = "^60.2"
redis = "^5.0.1"
django-rosetta = "^0.10.0"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.27.1"
psycopg = { extras = ["binary"], version = "^3.1.19" }
django-environ = "^0.11.2"
watchdog = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
